+++ src/main/java/org/netbeans/modules/hyperledger/cto/node/MembersFactory.java
@@ -27,8 +27,8 @@ import org.netbeans.modules.hyperledger.LookupContext;
 import org.netbeans.modules.hyperledger.cto.CtoResource;
 import org.netbeans.modules.hyperledger.cto.grammar.CtoLexer;
 import org.netbeans.modules.hyperledger.cto.grammar.CtoParser;
-import org.netbeans.modules.hyperledger.cto.grammar.ParserListener;
-import org.netbeans.modules.hyperledger.cto.grammar.ParserProvider;
+import org.netbeans.modules.hyperledger.cto.parser.ParserListener;
+import org.netbeans.modules.hyperledger.cto.parser.ParserProvider;
 import org.openide.filesystems.FileChangeAdapter;
 import org.openide.filesystems.FileEvent;
 import org.openide.filesystems.FileObject;
