+++ src/main/java/org/netbeans/modules/hyperledger/cto/parser/CtoProxyParser.java
@@ -23,7 +23,6 @@ import java.util.function.Function;
 import javax.swing.event.ChangeListener;
 import org.netbeans.modules.hyperledger.cto.CtoResource;
 import org.netbeans.modules.hyperledger.cto.grammar.CtoParser;
-import org.netbeans.modules.hyperledger.cto.grammar.ParserListener;
 import org.netbeans.modules.parsing.api.Snapshot;
 import org.netbeans.modules.parsing.api.Task;
 import org.netbeans.modules.parsing.spi.ParseException;
