+++ src/main/java/org/netbeans/modules/hyperledger/cto/parser/ParserProvider.java
@@ -16,37 +16,30 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.netbeans.modules.hyperledger.cto.grammar;
+package org.netbeans.modules.hyperledger.cto.parser;
 
-import org.antlr.v4.runtime.Vocabulary;
+import java.util.function.Function;
+import org.antlr.v4.runtime.CharStream;
+import org.antlr.v4.runtime.CharStreams;
+import org.antlr.v4.runtime.CommonTokenStream;
+import org.antlr.v4.runtime.Lexer;
+import org.antlr.v4.runtime.TokenStream;
+import org.netbeans.modules.hyperledger.cto.grammar.CtoLexer;
+import org.netbeans.modules.hyperledger.cto.grammar.CtoParser;
 
 /**
  *
  * @author mario.schroeder
  */
-public final class CtoVocabulary implements Vocabulary {
-
-    private final static Vocabulary VOCABULARY = CtoLexer.VOCABULARY;
-
-    @Override
-    public int getMaxTokenType() {
-        return VOCABULARY.getMaxTokenType();
-    }
+public enum ParserProvider implements Function<String, CtoParser> {
+    
+    INSTANCE;
 
     @Override
-    public String getLiteralName(int tokenType) {
-        return VOCABULARY.getLiteralName(tokenType);
+    public CtoParser apply(String text) {
+        CharStream input = CharStreams.fromString(text);
+        Lexer lexer = new CtoLexer(input);
+        TokenStream tokenStream = new CommonTokenStream(lexer);
+        return new CtoParser(tokenStream);
     }
-
-    @Override
-    public String getSymbolicName(int tokenType) {
-        return VOCABULARY.getSymbolicName(tokenType);
-    }
-
-    @Override
-    public String getDisplayName(int tokenType) {
-        String name = VOCABULARY.getDisplayName(tokenType);
-        return name.replaceAll("^\\'|\\'$", "");
-    }
-
 }
