+++ src/test/java/org/netbeans/modules/hyperledger/cto/parser/CtoProxyParserTest.java
@@ -24,7 +24,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.netbeans.modules.hyperledger.cto.grammar.ParserProvider;
 import org.netbeans.modules.hyperledger.cto.parser.CtoProxyParser.CtoParserResult;
 import org.netbeans.modules.parsing.api.Snapshot;
 import org.netbeans.modules.parsing.api.Task;
